a function that returns the number of bits you would need to flip to get from one number to another

+++ 0x14-bit_manipulation/5-flip_bits.c
@@ -8,7 +8,7 @@
  */
 unsigned int flip_bits(unsigned long int n, unsigned long int m)
 {
-	unsigned int xor_result, bit_count;
+	unsigned long int xor_result, bit_count;
 
 	xor_result = n ^ m;
 	bit_count = 0;
